refactor RandomTSPMatrix test cases
